Modify: strictly follow MVC structure.
