--- controllers/user_cont.go
@@ -1,45 +1,0 @@
-package controller
-
-import (
-	"chronote/global"
-	"chronote/models"
-	"chronote/utils"
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-func Register(ctx *gin.Context) {
-	var user models.User
-
-	if err := ctx.ShouldBindJSON(&user); err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-
-	hashedPassword, err := utils.EncryptPassword(user.Password)
-
-	if err != nil {
-		ctx.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
-	}
-
-	user.Password = hashedPassword
-
-	if err := global.Db.AutoMigrate(&user); err != nil {
-		ctx.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-
-	if err := global.Db.Create(&user).Error; err != nil {
-		ctx.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-}
